requirements.txt
absl-py
aiofiles==24.1.0
aiohappyeyeballs
aiohttp
aiosignal
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
async-timeout
attrs
beautifulsoup4==4.13.4
bidict==0.23.1
blinker
Bottleneck
Brotli
cachetools
certifi
cffi
chardet==5.2.0
charset-normalizer
click
colorama
contourpy
cryptography
cycler
docutils==0.21.2
exceptiongroup==1.3.0
fastapi==0.115.12
flatbuffers
fonttools
frozenlist
gast
google-auth
google-auth-oauthlib
google-pasta
grpcio
h11==0.16.0
h5py
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna
ifaddr==0.2.0
imageio
itsdangerous==2.2.0
Jinja2==3.1.6
joblib
keras
Keras-Preprocessing
kiwisolver
lazy_loader
libclang==18.1.1
Markdown
markdown2==2.5.3
MarkupSafe
matplotlib==3.10.0
mkl-service==2.4.0
mkl_fft
mkl_random
multidict
networkx
nicegui==2.17.0
numexpr
numpy
oauthlib
opencv-python==4.11.0.86
opt-einsum
orjson==3.10.18
packaging
pandas
pillow
ply==3.11
propcache
protobuf==3.19.6
pscript==0.7.7
pyasn1
pyasn1-modules==0.2.8
pycparser
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT
pyOpenSSL
pyparsing
PyQt5==5.15.10
PyQt5_sip
PySocks
python-dateutil
python-dotenv==1.1.0
python-engineio==4.12.1
python-multipart==0.0.20
python-socketio==5.13.0
pytz
PyYAML==6.0.2
reportlab==4.4.0
requests
requests-oauthlib
rsa
scikit-image
scikit-learn
scipy
simple-websocket==1.1.0
sip
six
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor
threadpoolctl
tifffile
tomli
tornado
typing-inspection==0.4.0
typing_extensions
tzdata
unicodedata2
urllib3
uvicorn==0.34.2
vbuild==0.8.2
wait_for2==0.3.2
watchfiles==1.0.5
websockets==15.0.1
Werkzeug
win-inet-pton
wrapt
wsproto==1.2.0
yarl
tf2onnx
screeninfo
